requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
Flask-MQTT==1.1.1
frozenlist==1.3.1
future==0.18.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meraki==1.25.0
multidict==6.0.2
paho-mqtt==1.6.1
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
webexteamssdk==1.6.1
Werkzeug==2.1.2
yarl==1.8.1
zipp==3.8.0